transfer logic from helper to lib
