+++ app/controllers/application_controller.rb
@@ -1,2 +1,4 @@
 class ApplicationController < ActionController::Base
+  require 'HelperModule.rb'
+  include HelperModule
 end
