+++ app/controllers/events_controller.rb
@@ -1,6 +1,4 @@
 class EventsController < ApplicationController
-  include SessionsHelper
-  include EventsHelper
 
   def new
     redirect_to root_path if current_user.nil?
