+++ app/helpers/events_helper.rb
@@ -1,5 +1,3 @@
 module EventsHelper
-  def creator(event)
-    User.find(event.creator_id)
-  end
+ 
 end
